Initial support for Darwin.
Basically the same as BSD, except that it lacks rfork, which
makes it impossible to initialize watchers before daemonizing.

* configure.ac: Check for rfork, define the DIRCOND_FORK conditional
if present.
* src/Makefile.am [DIRCOND_KQUEUE][DIRCOND_RFORK]: Add rdaemon.c
to the sources.
* src/dircond.c: Define INIT_EARLY if it is safe to initialize
the event system before forking.  Currently the only supported
system that's not capable of that is Darwin.
(main): Call setup_watchers conditionally, before or after the
call to daemon.
Move call to evsys_init to setup_watchers.
Reorganize the main loop.
* src/dircond.h (event_mask_init): Change signature. The third
argument specifies which mask bits we're interested in.
All uses changed.
* src/event.c: Ditto.
* src/watcher.c (setup_watchers): Call evsys_init.