Generate genev_create event for subdirectories.
A genev_create event is generated for all files and directories below the
newly created one, if required by the configuration.  At the same time, new
watchers are installed.  This is illustrated by the testcase "createrec.at",
which version 4.0 wouldn't pass, because it incorrectly assumed that a
notification would arrive for each subdirectory or subfile, once the watcher
is installed for the parent directory.

* src/watcher.c (subwatcher_create): Return integer.
Take additional argument (notify).  If it is true, register
watchers for the subdirectories.
(deliver_ev_create): New function.
(watch_subdirs): Return number of watchers installed or -1
on error.
(watch_pathname): Remove. All callers use subwatcher_create
instead.
* tests/Makefile.am: Add new test.
* tests/testsuite.at: Likewise.
* tests/attrib.at: Use pwd -P, instead of plain pwd.
* tests/cmdexp.at: Likewise.
* tests/create.at: Likewise.
* tests/delete.at: Likewise.
* tests/env00.at: Likewise.
* tests/env01.at: Likewise.
* tests/env02.at: Likewise.
* tests/env03.at: Likewise.
* tests/write.at: Likewise.